requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
tqdm>=4.66.4
pandas>=2.2.2
datasets>=2.20.0
chromadb>=0.5.3
sentence-transformers>=3.0.1
numpy>=1.26.4
scikit-learn>=1.4.2
streamlit>=1.36.0
